Implement context input
If given and un-empty, the `context` string is written as the body of
the lock object on S3. When a lock is acquired or found blocking, that
context is read and included in the details show.

This can be used to supply a sort of "about me" value at acquire time,
so that we can display a sort of "held by" when waiting.

By default, this input is the workflow and run number, which means
waiting on a lock in one run of a workflow will now show you which other
run of that workflow is currently holding it.