Retain place in line when waiting on a lock
Prior to this change a new object key was created each time we attempted
to acquire it. This meant that someone else could jump in and take our
place in line, since our created-at was updating each loop.

There are trade-offs either way, but this has happened many times now
where a process is waiting on a lock, then just before that lock expires
a new one cuts in line and we start waiting the full duration again.

This change sets created-at up front and then re-uses it on each loop.
It still extends the expiry each time, so we will not expire early, but
it ensures that, once we start waiting, no other lock that begins
waiting after us could show up as older before we do.

Ultimately, this should result in fairer outcomes.